Current value is pushed onto stack in all cases
The way I had written the state machine missed a few spots where I
should have pushed the current value onto the stack, basically any time
when we didn't end a value with a non-newline whitespace.

The new code is quite a bit simpler, and always pushes the current value
onto the stack whenever we read a character that isn't part of the
current number.